d3v4: move data fns to DataParse.js
